Move handle-based static field get to top level class
